requirements.txt
# Core dependencies
streamlit==1.29.0
langchain==0.1.0
langchain-groq==0.0.1
groq==0.4.1
python-dotenv==1.0.0
yagmail==0.15.293
apscheduler==3.10.4

# Additional utilities
pandas==2.1.4
requests==2.31.0
pydantic==2.5.2
typing-extensions==4.8.0

# Development and testing
pytest==7.4.3
black==23.12.0
flake8==6.1.0
